+++ 0x08-recursion/main.h
@@ -12,6 +12,6 @@ int _sqrt(int n, int i);
 int is_prime_number(int n);
 int _is_prime_number(int n, int i);
 int is_palindrome(char *s);
-int _is_palindrome(char *s, int first, int last);
+int _is_palindrome(char *s, unsigned long int first, unsigned long int last);
 
 #endif
